Split SQS from fs2-aws module
Separate SQS from fs2-aws module to reduce dependency footprint.

- Moved SQS code to `fs2-aws-sqs` and `fs2-aws-sqs-testkit`
- SQS Testkit classes now resides in `fs2.aws.sqs.testkit`
- Fix SqsSpec that isn't actually being run
