+++ build.sbt
@@ -121,22 +121,38 @@ lazy val `fs2-aws` = (project in file("fs2-aws"))
   .settings(
     name := "fs2-aws",
     libraryDependencies ++= Seq(
-      "co.fs2"                  %% "fs2-core"                     % V.Fs2,
-      "co.fs2"                  %% "fs2-io"                       % V.Fs2,
-      "com.amazonaws"           % "aws-java-sdk-kinesis"          % V.AwsSdk,
-      "com.amazonaws"           % "aws-java-sdk-s3"               % V.AwsSdk,
-      "com.amazonaws"           % "aws-java-sdk-sqs"              % V.AwsSdk,
-      "com.amazonaws"           % "amazon-kinesis-producer"       % "0.14.0",
-      "software.amazon.kinesis" % "amazon-kinesis-client"         % "2.2.11",
-      "org.mockito"             % "mockito-core"                  % V.MockitoCore % Test,
-      "software.amazon.awssdk"  % "sts"                           % "2.13.50",
-      "org.scalatest"           %% "scalatest"                    % V.ScalaTest % Test,
-      "org.mockito"             %% "mockito-scala-scalatest"      % V.MockitoScalaTest % Test,
-      "com.amazonaws"           % "aws-java-sdk-sqs"              % V.AwsSdk excludeAll ("commons-logging", "commons-logging"),
-      "com.amazonaws"           % "amazon-sqs-java-messaging-lib" % "1.0.8" excludeAll ("commons-logging", "commons-logging"),
-      "eu.timepit"              %% "refined"                      % V.Refined
+      "co.fs2"                  %% "fs2-core"                % V.Fs2,
+      "co.fs2"                  %% "fs2-io"                  % V.Fs2,
+      "com.amazonaws"           % "aws-java-sdk-kinesis"     % V.AwsSdk,
+      "com.amazonaws"           % "aws-java-sdk-s3"          % V.AwsSdk,
+      "com.amazonaws"           % "amazon-kinesis-producer"  % "0.14.0",
+      "software.amazon.kinesis" % "amazon-kinesis-client"    % "2.2.11",
+      "org.mockito"             % "mockito-core"             % V.MockitoCore % Test,
+      "software.amazon.awssdk"  % "sts"                      % "2.13.50",
+      "org.scalatest"           %% "scalatest"               % V.ScalaTest % Test,
+      "org.mockito"             %% "mockito-scala-scalatest" % V.MockitoScalaTest % Test,
+      "eu.timepit"              %% "refined"                 % V.Refined
     ),
     coverageMinimum       := 40,
+    coverageFailOnMinimum := true
+  )
+  .settings(commonSettings)
+  .settings(scalacOptions := commonOptions(scalaVersion.value))
+
+lazy val `fs2-aws-sqs` = (project in file("fs2-aws-sqs"))
+  .settings(
+    name := "fs2-aws",
+    libraryDependencies ++= Seq(
+      "co.fs2"        %% "fs2-core"                     % V.Fs2,
+      "co.fs2"        %% "fs2-io"                       % V.Fs2,
+      "com.amazonaws" % "aws-java-sdk-sqs"              % V.AwsSdk excludeAll ("commons-logging", "commons-logging"),
+      "com.amazonaws" % "amazon-sqs-java-messaging-lib" % "1.0.8" excludeAll ("commons-logging", "commons-logging"),
+      "org.mockito"   % "mockito-core"                  % V.MockitoCore % Test,
+      "org.scalatest" %% "scalatest"                    % V.ScalaTest % Test,
+      "org.mockito"   %% "mockito-scala-scalatest"      % V.MockitoScalaTest % Test,
+      "eu.timepit"    %% "refined"                      % V.Refined
+    ),
+    coverageMinimum       := 55.80,
     coverageFailOnMinimum := true
   )
   .settings(commonSettings)
@@ -151,6 +167,19 @@ lazy val `fs2-aws-testkit` = (project in file("fs2-aws-testkit"))
       "io.circe"      %% "circe-generic"           % V.Circe,
       "io.circe"      %% "circe-generic-extras"    % V.Circe,
       "io.circe"      %% "circe-parser"            % V.Circe,
+      "org.scalatest" %% "scalatest"               % V.ScalaTest,
+      "org.mockito"   % "mockito-core"             % V.MockitoCore,
+      "org.mockito"   %% "mockito-scala-scalatest" % V.MockitoScalaTest
+    )
+  )
+  .settings(commonSettings)
+  .settings(scalacOptions := commonOptions(scalaVersion.value))
+
+lazy val `fs2-aws-sqs-testkit` = (project in file("fs2-aws-sqs-testkit"))
+  .dependsOn(`fs2-aws-sqs`)
+  .settings(
+    name := "fs2-aws-testkit",
+    libraryDependencies ++= Seq(
       "org.scalatest" %% "scalatest"               % V.ScalaTest,
       "org.mockito"   % "mockito-core"             % V.MockitoCore,
       "org.mockito"   %% "mockito-scala-scalatest" % V.MockitoScalaTest
