+++ fs2-aws-sqs-testkit/src/main/scala/fs2/aws/sqs/testkit/TestSqsConsumerBuilder.scala
@@ -1,10 +1,10 @@
-package fs2.aws.testkit
+package fs2.aws.sqs.testkit
 import cats.effect.Effect
 import com.amazon.sqs.javamessaging.SQSConnection
-import fs2.aws.testkit.TestSqsConsumerBuilder.TestSQSConsumer
+import fs2.aws.sqs.testkit.TestSqsConsumerBuilder.TestSQSConsumer
 import fs2.aws.sqs.{ ConsumerBuilder, SQSConsumer }
 import javax.jms.MessageListener
-import org.mockito.scalatest.MockitoSugar
+import org.mockito.MockitoSugar
 
 class TestSqsConsumerBuilder[F[_]: Effect] extends ConsumerBuilder[F] {
   override def start: F[SQSConsumer] =
