+++ fs2-aws-sqs/src/main/scala/fs2/aws/sqs/SqsConfig.scala
