+++ fs2-aws-sqs/src/main/scala/fs2/aws/sqs/package.scala
